requirements.txt
# requirements.txt
playwright==1.49.0
beautifulsoup4==4.12.3
requests==2.32.3
gspread==6.1.4
google-auth==2.37.0
google-auth-oauthlib==1.2.1
google-genai>=1.0.0
python-dotenv==1.0.1
fastapi==0.115.6
uvicorn==0.34.0
apscheduler==3.10.4
httpx==0.28.1
python-telegram-bot==21.9
pytest==8.3.4
